+++ modules/SofaSimpleFem/TetrahedronFEMForceField.h
@@ -191,7 +191,7 @@ public:
 
     void getRotations(VecReal& vecR)
     {
-        unsigned int nbdof = this->mstate->getSize();
+        unsigned int nbdof = (unsigned int)this->mstate->getSize();
         for (unsigned int i=0; i<nbdof; ++i)
         {
 
@@ -201,7 +201,7 @@ public:
 
     void getRotations(defaulttype::BaseMatrix * rotations,int offset = 0)
     {
-        unsigned int nbdof = this->mstate->getSize();
+        unsigned int nbdof = (unsigned int)this->mstate->getSize();
 
         if (component::linearsolver::RotationMatrix<float> * diag = dynamic_cast<component::linearsolver::RotationMatrix<float> *>(rotations))
         {
